+++ src/main.js
@@ -10,15 +10,7 @@ Vue.config.productionTip = false
 
 /* eslint-disable */
 // For Firebase JS SDK v7.20.0 and later, measurementId is optional
-const config = {
-  apiKey: "AIzaSyDeNuzV_lTclsv8iYhfH-K_sV9cNYOHUs4",
-  authDomain: "tipping-app-be890.firebaseapp.com",
-  projectId: "tipping-app-be890",
-  storageBucket: "tipping-app-be890.appspot.com",
-  messagingSenderId: "523721486484",
-  appId: "1:523721486484:web:0db73f6f5d192b894f8481",
-  measurementId: "G-TF5NZJT4LR"
-};
+import { config } from '@/utilities/firebase.config';
 
 firebase.initializeApp(config);
