+++ src/router/index.js
@@ -1,17 +1,17 @@
 import Vue from 'vue'
 import Router from 'vue-router'
-import HelloWorld from '@/components/HelloWorld'
+import DashBoard from '@/components/DashBoard'
 import Signup from '@/components/Signup'
 import Signin from '@/components/Signin'
 
-Vue.use(Router)
+const router = Vue.use(Router)
 
 export default new Router({
   routes: [
     {
       path: '/',
-      name: 'HelloWorld',
-      component: HelloWorld
+      name: 'DashBoard',
+      component: DashBoard
     },
     {
       path: '/signup',
@@ -25,3 +25,5 @@ export default new Router({
     }
   ]
 })
+
+export { router }
