+++ src/store/index.js
@@ -3,6 +3,9 @@
 
 import Vue from 'vue'
 import Vuex from 'vuex'
+import firebase from 'firebase'
+import router from '@/router'
+
 
 Vue.use(Vuex)
 
@@ -10,7 +13,21 @@ const store = new Vuex.Store({
     state: {
     },
     mutations: {
+        signUp: function (state) {
+            firebase.auth().createUserWithEmailAndPassword(state.email, state.password)
+                .then(() => {
+                    firebase.auth().currentUser.updateProfile({
+                    displayName: state.username,
+                })
+                .then(() => {
+                    router.push('/')
+                })
+                })
+                .catch(error => {
+                    alert(error.message)
+                })
+        },
     },
-  });
+});
 
 export default store
